+++ lib/mlx/io.rb
@@ -8,3 +8,4 @@ end
 require_relative "io/safetensors"
 require_relative "io/hub"
 require_relative "io/huggingface"
+require_relative "io/adapter"